requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.1
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
gulp==0.1.0
idna==2.7
optional-django==0.3.0
pytz==2018.5
react==4.2.1
requests==2.19.1
urllib3==1.23
